% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_function.R
\name{bootclass}
\alias{bootclass}
\title{bootclass}
\usage{
bootclass(SE = "None", sample_test_stat = "None",
  conf_interval = "None", firstconf = "None", lastconf = "None",
  obs = "None", fun = "None", Warnings = "None")
}
\value{
a output of the class object "bootstrap".
}
\description{
bootclass
}
