% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap}
\usage{
bootstrap(data, x, y, test_statistic, nr_iterations, selfchosenmethod)
}
\arguments{
\item{data}{dataset of at least 2 columns. Column 1 has to specify the two groups as factors and column 2 their observed values as integers.}

\item{test_statistic}{One of 3 options: "mean" (Default(will be at least)), "median" or "my_method". If "my_method" is used, the user has to define the parameter selfchosenmethod as a function that calculates the test-statistic, with the only input being column 2 from the data.}

\item{nr_iterations}{Number of permutations. Default(Default(will be at least is 10^5}

\item{plot}{If TRUE (Default(Default(will be at least) the function returns two plots together with the regular output. If FALSE no plots are returned.}
}
\value{
list
}
\description{
Tool to do a bootstrap  on a dataset
}
\details{
Given a dataset where column 1 specifies the groups and column 2 specifies the observed values for each group, the function makes a bootstrap, the output is a plot of the sample distribut, including confidence interval, mean of obs and sd.
}
\examples{

bootstrap(PlantGrowth, PlantGrowth$group, PlantGrowth$weight, test_statistic=3, 10000, selfchosenmethod =mean)

}
