% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_plots.R
\name{corr.plot.permutation}
\alias{corr.plot.permutation}
\title{Correltation matrix between original and permuted data}
\usage{
corr.plot.permutation(
  seurat_analysis,
  time.var = "time_num",
  perm.list = c("perm1", "perm2", "perm3", "perm4", "perm5"),
  numeric.function = function(x) {
     as.numeric(x)
 }
)
}
\arguments{
\item{seurat_analysis}{seurat object}

\item{time.var}{Name of time label column in the metadata of the seurat object}

\item{perm.list}{Name of all permutation columns in the metadata of the seurat object to be plotted}

\item{numeric.function}{function to generate numeric version of the time and permutation labels}
}
\value{
Correlation matrix plot
}
\description{
Correltation matrix between original and permuted data
}
\examples{
\dontrun{corr.plot.permutation(seurat_analysis, time.var = "time",
numeric.function = function(x){as.numeric(str_split(x, pattern =  "W", simplify = T)[,2])} )}
}
