% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_power.R
\name{detection_power_padj}
\alias{detection_power_padj}
\title{Detection power calculated from the Power DE call}
\usage{
detection_power_padj(DEG.power, padj = NULL)
}
\arguments{
\item{DEG.power}{Output of the Power DE call}

\item{padj}{Name of the adjusted pvalue column. If NULL the adjusted pvalue column is generated based on regular expression.}
}
\value{
Detection power calculated from the Power DE call
}
\description{
Detection power calculated from the Power DE call
}
\examples{
\dontrun{detection_power_padj(DEG.power)}
}
