% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_power.R
\name{generate_pseudobulk}
\alias{generate_pseudobulk}
\title{Generate pseudobulk based on donor and time label}
\usage{
generate_pseudobulk(
  seurat_analysis,
  gene_filter = seurat_analysis@assays$RNA@var.features,
  gene_filter_count = NULL,
  time.var = "time",
  assay = "RNA",
  design_matrix = ~time + donor,
  save_DDS = FALSE,
  outdir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_analysis}{seurat object}

\item{gene_filter}{Filter genes by list}

\item{gene_filter_count}{Filter genes by count}

\item{time.var}{Name of the time variable column}

\item{assay}{Name of the Assay in the seurat object}

\item{design_matrix}{Design matrix for DESeq2}

\item{save_DDS}{Logical. Save output as rds}

\item{outdir}{Output location for saved files}

\item{verbose}{Logical}
}
\value{
DESeq2 opbject
}
\description{
Generate pseudobulk based on donor and time label
}
\examples{
\dontrun{dds <- generate_pseudobulk(seurat_analysis)}
}
