% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_pseudobulk.R
\name{generate_pseudobulk_perm}
\alias{generate_pseudobulk_perm}
\title{Generate Pseudobulk from seurat object}
\usage{
generate_pseudobulk_perm(
  seurat_analysis,
  gene_filter = seurat_analysis@assays$RNA@var.features,
  gene_filter_count = NULL,
  perm = c("perm1", "perm2", "perm3", "perm4", "perm5"),
  assay = "RNA",
  design_matrix = ~time + donor,
  save_DDS = FALSE,
  outdir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{seurat_analysis}{seurat object}

\item{gene_filter}{Filter genes by list}

\item{gene_filter_count}{Filter genes by count}

\item{perm}{List of permutation names}

\item{assay}{Name of the Assay in the seurat object}

\item{design_matrix}{Design matrix for DESeq2}

\item{save_DDS}{Logical. Save output as rds}

\item{outdir}{Output location for saved files}

\item{verbose}{Logical}
}
\value{
DESeq2 pseudobulk object
}
\description{
Generate Pseudobulk from seurat object
}
\examples{
\dontrun{dds_perm <- generate_pseudobulk_perm(seurat_analysis = seurat_analysis)}
}
