% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_load_DEG.R
\name{load_DEG}
\alias{load_DEG}
\title{Load to DEG perm function}
\usage{
load_DEG(files = c(seurat_wilcox = "DEG_seurat_wilcox_perm.rds"))
}
\arguments{
\item{files}{List of files to load}
}
\value{
List of DE call data frames
}
\description{
Load to DEG perm function
}
\examples{
\dontrun{
DEG.perm <- load_DEG(c("seurat_wilcox" = "/data/Femmunity/analysis/laura/external_data/PALMO/v3/241212_DEG_seurat_wilcox_perm.rds"))
}
}
