% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{output}
\alias{output}
\title{output}
\usage{
output(data = data.frame(), null.value = c(),
  alternative = "two-sided", method = "Two-sided permutation test",
  estimate = c(), data.name = "", statistic = c(),
  parameters = c(), perm_statistics = 0, p.value = 0,
  fun = "No function specified", Warnings = "None")
}
\value{
a list of class object "permutation".
}
\description{
output
}
