% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_FPR.R
\name{perm_FPR_df}
\alias{perm_FPR_df}
\title{FPR data frame}
\usage{
perm_FPR_df(
  DEG.perm = list(seurat_wilcox = DEG.seurat.perm.null, pseudobulk_DESeq2_Wald =
    DEG.pseudobulk_Wald.perm.null),
  permutations = c("perm1", "perm2", "perm3", "perm4", "perm5"),
  permutation_col = "permutation_round",
  pvalue = NULL,
  padj = NULL
)
}
\arguments{
\item{DEG.perm}{Output of Permutation DE call or load DEG}

\item{permutations}{Name of all permutations in the metadata of the seurat object}

\item{permutation_col}{Name of the permutation column in the DEG data frame}

\item{pvalue}{Name of the pvalue column. If NULL the pvalue column is generated based on regular expression.}

\item{padj}{Name of the adjusted pvalue column. If NULL the adjusted pvalue column is generated based on regular expression.}
}
\value{
Data frame of FPRs
}
\description{
FPR data frame
}
\examples{
\dontrun{fpr_df <- perm_FPR_df(DEG.perm = DEG.perm)}
}
