% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_DE_call.R
\name{Perm_QQplot}
\alias{Perm_QQplot}
\title{QQ plot of pvalues}
\usage{
Perm_QQplot(
  DEG.perm = list(seurat_wilcox = DEG.seurat.perm.null, pseudobulk_DESeq2_Wald =
    DEG.pseudobulk_Wald.perm.null),
  pvalue = NULL,
  color = c(tradeseq_time = "#c6dbef", tradeseq_slingshot = "#2171b5",
    tradeseq_psupertime = "#032f5f", pseudobulk_DESeq2_Wald = "#A65628",
    pseudobulk_DESeq2_LRT = "#ef7c00", seurat_wilcox = "#e7bd00", seurat_MAST =
    "#fb6a4a", MAST_RE = "#cb181d", seurat_MAST_latent = "#a50f15", MAST_longitudinal =
    "#a1d99b", MAST_longitudinal_RE = "#006d2c", MAST_longitudinal_PID = "#00441b",
    linear_TDE = "#ffb3ff", mixed_TDE = "#984EA3")
)
}
\arguments{
\item{DEG.perm}{Output of Permutation DE call or load DEG}

\item{pvalue}{Name of the pvalue column. If NULL the pvalue column is generated based on regular expression.}

\item{color}{color vector}
}
\value{
QQ plot of pvalues
}
\description{
QQ plot of pvalues
}
\examples{
\dontrun{Perm_QQplot(DEG.perm = DEG.perm)}
}
