% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Permutation-package.R
\docType{package}
\name{Permutation-package}
\alias{Permutation}
\alias{Permutation-package}
\title{Permutation: Permutation workflow to calculate FDR in a given dataset}
\description{
Permutation workflow to calculate FDR in a given dataset.
}
\author{
\strong{Maintainer}: Laura Lang \email{llang@uni-bonn.de}

}
\keyword{internal}
