% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{permutation}
\alias{permutation}
\title{Generate permutation of time labels}
\usage{
permutation(
  seurat_analysis,
  method = "longitudinal",
  donor.var = "PID",
  time.var = "time",
  remove.previous = T,
  seed = c(360, 90, 477, 26, 551),
  perm_prefix = NULL,
  numeric = "default",
  numeric.function = NULL
)
}
\arguments{
\item{seurat_analysis}{Seurat object}

\item{method}{Can be "longitudinal", "cross-sectional" that preserve the original structure of the data or "random" that might distort the data structure}

\item{donor.var}{Variable with unique donor identifiers}

\item{time.var}{Variable with time labels}

\item{remove.previous}{Remove previously generated permutations that contain the "perm" in the name}

\item{seed}{Set as many seeds as permutation rounds}

\item{perm_prefix}{Add a prefix to the permutation output column name}

\item{numeric}{Generates numeric version of the time labels needed for some DE tools as input. If default the output is generated by as.numeric() if custom then specify the function to get numeric output in numeric.function}

\item{numeric.function}{Specify the function to get numeric output of the time labels}
}
\value{
seurat object with the permuted time labels in the meta data
}
\description{
Generate permutation of time labels
}
\examples{
\dontrun{seurat_analysis <- permutation(seurat_analysis, method = "donor_longitudinal", numeric = "custom",
numeric.function = function(x){as.numeric(str_split(x, pattern =  "W", simplify = T)[,2])})}

}
