% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_cor.R
\name{permutation_cor}
\alias{permutation_cor}
\title{permutation_cor}
\usage{
permutation_cor(x, y, iterations = 10^4, method = "pearson", seed = 4)
}
\arguments{
\item{x}{One of the two vectors/data frame columns to be correlated}

\item{y}{The second of the two vectors/data frame columns to be correlated}

\item{iterations}{Number of permutations (default = 10000)}

\item{method}{Correlation method (default = 'pearson')}

\item{seed}{Set for reproducibility (default = 4)}
}
\value{
Results reported in the console; plot generated in the viewer
}
\description{
Runs a non-parametric permutation test of a correlation.
}
\details{
\code{permutation_cor()} will:
\itemize{
  \item Run a permutation test for a correlation and report a (two-sided) p-value
  \item Report a symmetry test for the null distribution generated from permuted data (\code{symmetry.test()} called from \code{lawstat})
  \item Produce a histogram of the generated null distribution and overlay the observed correlation on it (using \code{ggplot2})
}
Requires: \code{ggplot2} and \code{lawstat}
}
\examples{
set.seed(1969)
x <- rnorm(1000, 0, 1)
y <- x + runif(1000, 0, 10)
permutation_cor(x, y)

}
