% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_mdiff.R
\name{permutation_mdiff}
\alias{permutation_mdiff}
\title{permutation_mdiff}
\usage{
permutation_mdiff(
  groups,
  values,
  iterations = 10^4,
  seed = 4,
  downsample = FALSE
)
}
\arguments{
\item{groups}{The vector/data frame column containing the (two) group identifiers}

\item{values}{The vector/data frame column containing the values to be compared (numeric or numeric-coercible)}

\item{iterations}{Number of permutations (default = 10000)}

\item{seed}{Set for reproducibility (default = 4)}

\item{downsample}{If groups A and B are unbalanced, you can set the \code{downsample} argument to TRUE to use downsampling each permutation to generate permuted groups of size min(n_A, n_B); otherwise, \code{permutation_mdiff} will generate permuted groups of size n_A and n_B during the permutation procedure (default = FALSE)}
}
\value{
Results reported in the console; plot generated in the viewer
}
\description{
Runs a non-parametric permutation test of the mean difference between two groups.
}
\details{
\code{permutation_mdiff} will:
\itemize{
  \item Run a permutation test for the mean difference between two groups and report a two-sided p-value
  \item Report a symmetry test for the null distribution generated from permuted data (\code{symmetry.test} called from \code{lawstat})
  \item Produce a histogram of the generated null distribution and overlay the observed mean difference on it (using \code{ggplot2})
}
Requires: \code{ggplot2} and \code{lawstat}

Note: If groups A and B are unbalanced, \code{permutation_mdiff}'s default is to generate permuted groups of sizes n_A and n_B; you can, however, use a downsampling procedure to generate permuted groups of size min(n_A, n_B); see above.
}
\examples{
set.seed(1969)
groups <- c(rep('control', 500), rep('experimental', 495))
values <- c(runif(500, 0, 5), runif(490, 0.25, 5), rep(NA, 5))
permutation_mdiff(groups, values)
permutation_mdiff(groups, values, downsample = T)

}
