% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_t.R
\name{permutation_t}
\alias{permutation_t}
\title{permutation_t}
\usage{
permutation_t(groups, values, iterations = 10^4, seed = 4, one.sided = FALSE)
}
\arguments{
\item{groups}{The vector/data frame column containing the (two) group identifiers}

\item{values}{The vector/data frame column containing the values to be compared}

\item{iterations}{Number of permutations (default = 10000)}

\item{seed}{Set for reproducibility (default = 4)}
}
\value{
Results reported in the console; plot generated in the viewer
}
\description{
Runs a non-parametric permutation test of the mean difference between two groups.
}
\details{
\code{permutation_t()} will:
\itemize{
  \item Run a permutation test for the mean difference between two groups and report a (two-sided) p-value
  \item Report a symmetry test for the null distribution generated from permuted data (\code{symmetry.test()} called from \code{lawstat})
  \item Produce a histogram of the generated null distribution and overlay the observed mean difference on it (using \code{ggplot2})
}
Requires: \code{ggplot2} and \code{lawstat}
}
\examples{
set.seed(1969)
groups <- c(rep('group_a', 500), rep('group_b', 500))
values <- c(rnorm(500, 0, 1), rnorm(500, 0.2, 1))
dat <- data.frame(groups, values)
permutation_t(dat$groups, dat$values)

}
