% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_plots.R
\name{plot.donor.permutation}
\alias{plot.donor.permutation}
\title{Plot cell numbers per donor}
\usage{
plot.donor.permutation(
  seurat_analysis,
  donor.var = "PID",
  time.var = c("time", "perm1", "perm2", "perm3", "perm4", "perm5"),
  color = brewer.pal(n = 9, name = "Set1"),
  plots = c("relative", "absolute")
)
}
\arguments{
\item{seurat_analysis}{seurat object}

\item{donor.var}{Name of donor ID column in the metadata of the seurat object}

\item{time.var}{Name of time label column in the metadata of the seurat object}

\item{color}{color vector}

\item{plots}{Choose between "relative" for relative cell numbers per donor, "absolute" for absolute cell numbers per donor}
}
\value{
Plot of cell numbers per donor
}
\description{
Plot cell numbers per donor
}
\examples{
\dontrun{plot.donor.permutation(seurat_analysis, color = color_time)}
}
