% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_plots.R
\name{plot.pseudotime.permutation}
\alias{plot.pseudotime.permutation}
\title{Plot time and donor variable over pseudotime per permutation}
\usage{
plot.pseudotime.permutation(
  seurat_analysis,
  donor.var = "PID",
  time.var = "time",
  perm = c("pseudotime", "pseudotime_perm1", "pseudotime_perm2", "pseudotime_perm3",
    "pseudotime_perm4", "pseudotime_perm5"),
  color = c(time = brewer.pal(n = 9, name = "Set1"), donor = brewer.pal(n = 9, name =
    "Set2")),
  plots = c("time", "donor")
)
}
\arguments{
\item{seurat_analysis}{seurat object}

\item{donor.var}{Name of donor ID column in the metadata of the seurat object}

\item{time.var}{Name of original time label column in the metadata of the seurat object}

\item{perm}{Name of pseudotime and permutation columns to plot}

\item{color}{color vector for time and donor}

\item{plots}{Choose between "time" and "donor" for time or donor labels, respectively}
}
\description{
Plot time and donor variable over pseudotime per permutation
}
