% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_DE_call.R
\name{plot_DEG}
\alias{plot_DEG}
\title{Plot number of DEG genes per permutation}
\usage{
plot_DEG(DEG.perm = DEG.perm, alpha = 0.05)
}
\arguments{
\item{DEG.perm}{Output of Permutation DE call or load DEG}

\item{alpha}{Significance cut off}
}
\value{
Plot number of DEG genes per permutation
}
\description{
Plot number of DEG genes per permutation
}
\examples{
\dontrun{plot_DEG(DEG.perm = DEG.perm)}
}
