% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_DE_call.R
\name{plot_DEG_power}
\alias{plot_DEG_power}
\title{Plot number of DEG genes  for power DE call}
\usage{
plot_DEG_power(DEG.power = DEG.power, alpha = 0.05)
}
\arguments{
\item{DEG.power}{Output of Permutation DE call or load DEG}

\item{alpha}{Significance cut off}
}
\value{
Plot number of DEG genes  for power DE call
}
\description{
Plot number of DEG genes  for power DE call
}
\examples{
\dontrun{plot_DEG_power(DEG.perm = DEG.perm)}
}
