% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_power.R
\name{plot_detection_power}
\alias{plot_detection_power}
\title{Plot detection power}
\usage{
plot_detection_power(avg_power, color = color_method)
}
\arguments{
\item{avg_power}{Output from detection_power_padj}

\item{color}{color vector}
}
\value{
Plot detection power
}
\description{
Plot detection power
}
\examples{
\dontrun{plot_detection_power(avg_power)}
}
