% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_FPR.R
\name{plot_false_positives}
\alias{plot_false_positives}
\title{Plot false positives}
\usage{
plot_false_positives(fpr_df, color = color_method)
}
\arguments{
\item{fpr_df}{Output from perm_fpr_df}

\item{color}{color vector}
}
\value{
Plot of false positives
}
\description{
Plot false positives
}
\examples{
\dontrun{plot_false_positives(fpr_df)}
}
