% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_FPR.R
\name{plot_FPR_box}
\alias{plot_FPR_box}
\title{FPR boxplot}
\usage{
plot_FPR_box(fpr_df, color = color_method)
}
\arguments{
\item{fpr_df}{Output from perm_fpr_df}

\item{color}{color vector}
}
\value{
Boxplot of FPR
}
\description{
FPR boxplot
}
\examples{
\dontrun{plot_FPR_box(fpr_df)}
}
