% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_FPR.R
\name{plot_FPR_point}
\alias{plot_FPR_point}
\title{FPR scatterplot}
\usage{
plot_FPR_point(fpr_df, color = color_method)
}
\arguments{
\item{fpr_df}{Output from perm_fpr_df}

\item{color}{color vector}
}
\value{
Scatter plot of FPR values
}
\description{
FPR scatterplot
}
\examples{
\dontrun{plot_FPR_point(fpr_df)}
}
