% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_power.R
\name{Power_DEcall}
\alias{Power_DEcall}
\title{Power DE call}
\usage{
Power_DEcall(
  seurat_analysis,
  method = c("seurat_wilcox", "seurat_MAST", "seurat_MAST_latent"),
  donor.var = "PID",
  time.var = "time",
  comparison.group1 = c("V2", "V3", "V4"),
  comparison.group2 = rep("V1", 3),
  seurat.test.use = list(),
  min.pt = 0.1,
  gene_filter = seurat_analysis@assays$RNA@var.features,
  tde_method = c(linear = "cell", mixed = "pseudocell"),
  cellWeights = NULL,
  nknots = 6,
  save_DEG = TRUE,
  save_model = FALSE,
  outdir = NULL,
  verbose = TRUE,
  ncores = 1
)
}
\arguments{
\item{seurat_analysis}{Seurat object}

\item{method}{List of available options c("seurat_wilcox",
"linear_TDE",
"mixed_TDE",
"seurat_MAST",
"seurat_MAST_latent",
"MAST_longitudinal",
"MAST_longitudinal_RE",
"MAST_longitudinal_fixed")}

\item{donor.var}{Name of donor ID column in the metadata of the seurat object}

\item{time.var}{Name of time column in the metadata of the seurat object}

\item{comparison.group1}{Define pair-wise comparisons}

\item{comparison.group2}{Define pair-wise comparison control groups}

\item{seurat.test.use}{Passed to seurat FindMarkers test.use}

\item{min.pt}{Min percent filter}

\item{gene_filter}{Filter genes before model fitting}

\item{tde_method}{Can be "cell" or "pseudocell". See TDESeq for more information}

\item{cellWeights}{Cellweights from slingshot trajectory passed to TradeSeq FitGAM}

\item{nknots}{TradeSeq FitGAM nknots argument}

\item{save_DEG}{Logical. Save the output of the DE call as RDS}

\item{save_model}{Logical. Save the fitted models as RDS}

\item{outdir}{Define location where the output files are saved}

\item{verbose}{Logical}

\item{ncores}{Numbers of cores used}
}
\value{
List of DE output data frames
}
\description{
Power DE call
}
\examples{
 \dontrun{
DEG.power <- Power_DEcall(seurat_analysis = seurat_analysis, method =
c("seurat_wilcox","linear_TDE", "mixed_TDE","seurat_MAST","seurat_MAST_latent",
"MAST_longitudinal","MAST_longitudinal_RE","MAST_longitudinal_fixed"),comparison.group1 = c("W3", "W4", "W5", "W6","W7"),
comparison.group2 = rep("W2", 5),save_DEG = T, ncores = 5, gene_filter = NULL, tde_method = "pseudocell", outdir = outdir)
}
}
