% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_power.R
\name{Power_DEcall_pseudobulk}
\alias{Power_DEcall_pseudobulk}
\title{Power DE call pseudobulk}
\usage{
Power_DEcall_pseudobulk(
  dds,
  method = c("pseudobulk_DESeq2_Wald", "pseudobulk_DESeq2_LRT"),
  comparison.group1 = c("V2", "V3", "V4"),
  comparison.group2 = rep("V1", 3),
  design_matrix = ~time + donor,
  design_matrix_reduced = ~donor,
  time.var = "time",
  donor.var = "PID",
  save_DEG = TRUE,
  outdir = NULL,
  verbose = TRUE,
  ncores = 1,
  alpha = 0.05,
  pAdjustMethod = "BH",
  independentFiltering = TRUE,
  shrinkage = TRUE,
  shrinkType = "normal",
  multiple_testing = "IHW"
)
}
\arguments{
\item{dds}{Output from generate_pseudobulk}

\item{method}{"Wald" or "LRT"}

\item{comparison.group1}{Define pair-wise comparisons}

\item{comparison.group2}{Define pair-wise comparison control groups}

\item{design_matrix}{Design matrix for the DE call}

\item{design_matrix_reduced}{reduced design matrix}

\item{time.var}{Name of time column in the metadata of the seurat object}

\item{donor.var}{Name of donor ID column in the metadata of the seurat object}

\item{save_DEG}{Logical. Save the output of the DE call as RDS}

\item{outdir}{Define location where the output files are saved}

\item{verbose}{Logical}

\item{ncores}{Numbers of cores used}

\item{alpha}{Significance Threshold}

\item{pAdjustMethod}{Pvalue adjustment method}

\item{independentFiltering}{DE call parameter}

\item{shrinkage}{DE call parameter}

\item{shrinkType}{DE call parameter}

\item{multiple_testing}{DE call parameter}
}
\value{
Power DE call pseudobulk
}
\description{
Power DE call pseudobulk
}
\examples{
\dontrun{DEG.power <- Power_DEcall_pseudobulk(dds, method = c("pseudobulk_DESeq2_Wald","pseudobulk_DESeq2_LRT"),
comparison.group1 = c("W3", "W4", "W5", "W6","W7"),comparison.group2 = rep("W2", 5),save_DEG = T, outdir = outdir,
shrinkType = "apeglm")}
}
