% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_function.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap}
\usage{
bootstrap(observations, method = "mean", nboot = 10000)
}
\arguments{
\item{observations}{the column of  the dataframe with the observations}

\item{method}{One of 3 options: "mean" (Defaul), "median" or "selfchoosenmethod". If "selfchoosenmethod" is used, the user has to define the parameter selfchosenmethod as a function that calculates the test-statistic, with the only input being column 2 from the data.}

\item{nboot}{Number of iterations, Default(10000)}
}
\value{
list of class bootstrap
}
\description{
Tool to do a bootstrap  on a dataset
}
\details{
Given a dataset where column 1 specifies the groups and column 2 specifies the observed values for each group, the function makes a bootstrap, the output is a plot of the sample distribut, including confidence interval, mean of obs and sd.
}
\examples{

bootstrap(PlantGrowth$weight, method="mean", nboot = 10000)

summary(bootstrap)

plot(bootstrap, bins = 30)

}
