% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.R
\name{output}
\alias{output}
\title{output}
\usage{
output(data = tibble(), Method = "Unknown", fun = "mean",
  obs_Test_stat = 0, perm_test_values = 0, perm_p_value = 0,
  Warnings = "None")
}
\value{
a list of class object "permutation".
}
\description{
output
}
