\name{Sampling}
\alias{Sampling}

\title{Sampling}
\description{
  A package for doing permutations and bootstraping easily in R.
  This package provide a means for hypothesis testing when the data does not follow the assumtions of parametric test.
  A permutation test is done by doing resampling of the observed data without replacement,
  thereby creating the null-distribution for the association between two variables.
  Bootstraping is done by doing resampling of the observed data woth replacement,
  thereby creating the sampling distribution of a desirered estimate,
  e.g the difference in mean. From the Sampling distribution,
  the standart error and confidence intervals are calculated.
}

\author{
  Lotte Hindhede  & Cornelia Buus
}
